-popup done

+++ src/app/components/list-restaurant/list-restaurant.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonService } from 'src/app/services/common.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-restaurant',
@@ -16,27 +17,41 @@ export class ListRestaurantComponent implements OnInit {
 
   ngOnInit(): void {
     this.updateRestrauList();
+
+    this.commonService.isLoader();
   }
 
-  // updateList(data: any,id: number){
-  // console.log("🚀 ~ file: list-restaurant.component.ts ~ line 24 ~ ListRestaurantComponent ~ updateList ~ id", id)
-  // console.log("🚀 ~ file: list-restaurant.component.ts ~ line 24 ~ ListRestaurantComponent ~ updateList ~ data", data)
-
-  // }
-
-  updateRestrauList(){
+  updateRestrauList() {
     this.commonService.getRestrauList().subscribe(res => {
       console.log("🚀 ~ file: list-restaurant.component.ts ~ line 29 ~ ListRestaurantComponent ~ this.commonService.getRestrauList ~ res", res)
       this.restrauList = res;
     })
   }
 
-  onRemove(id){
-    this.restrauList.splice(id.id,-1);
-    this.commonService.removeRestrauDetails(id).subscribe(el => {
-    console.log("🚀 ~ file: list-restaurant.component.ts ~ line 26 ~ ListRestaurantComponent ~ this.commonService.removeRestrauDetails ~ el", el);
-    this.updateRestrauList();
+  onRemove(id) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+        this.restrauList.splice(id.id, -1);
+        this.commonService.removeRestrauDetails(id).subscribe(el => {
+          console.log("🚀 ~ file: list-restaurant.component.ts ~ line 26 ~ ListRestaurantComponent ~ this.commonService.removeRestrauDetails ~ el", el);
+          this.updateRestrauList();
+        })
+      }
     })
+
   }
 
   // deleteResto(resto){
